refactor: moved Service into handlers as well as it seems the place where every other service is
