+++ fhir-test-services/fhir-itb-service/src/main/java/eu/europa/ec/fhir/handlers/FhirProxyController.java
@@ -1,6 +1,5 @@
 package eu.europa.ec.fhir.handlers;
 
-import eu.europa.ec.fhir.FhirProxyService;
 import jakarta.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
